requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.0
exceptiongroup==1.1.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
install==1.3.5
loguru==0.7.0
multidict==6.0.4
numpy==1.25.1
openai==0.27.8
pinecone-client==2.2.2
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
